requirements.txt
Django~=5.1.7
djangorestframework~=3.15.2
psycopg2-binary~=2.9.10
django-debug-toolbar==3.4.0
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.22.1
Pillow==9.1.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
psycopg2-binary==2.9.8
python-dotenv